refactor: extract renderLabsSection
